backend/requirements.txt
Flask
Flask-CORS # For handling Cross-Origin Resource Sharing
Werkzeug # For secure_filename and other utilities
google-generativeai>=0.3.0 # Google's Gemini API
